+++ src/Servers/OSX.php
@@ -27,13 +27,18 @@ final class OSX implements Server
     private LoadAverageFacade $loadAverage;
     private UnixDisk $disk;
 
-    public function __construct(Clock $clock, Control $control, EnvironmentPath $path)
+    private function __construct(Clock $clock, Control $control, EnvironmentPath $path)
     {
         $this->cpu = new CpuFacade($control->processes());
         $this->memory = new MemoryFacade($control->processes(), $path);
-        $this->processes = new UnixProcesses($clock, $control->processes());
+        $this->processes = UnixProcesses::of($clock, $control->processes());
         $this->loadAverage = new LoadAverageFacade;
-        $this->disk = new UnixDisk($control->processes());
+        $this->disk = UnixDisk::of($control->processes());
+    }
+
+    public static function of(Clock $clock, Control $control, EnvironmentPath $path): self
+    {
+        return new self($clock, $control, $path);
     }
 
     #[\Override]